% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addLocation.R
\name{addLocation}
\alias{addLocation}
\title{Adds a new known location record to a table}
\usage{
addLocation(locationTbl = NULL, longitude = NULL, latitude = NULL,
  radius = NULL, stateDataset = "NaturalEarthAdm1", verbose = TRUE)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{longitude}{Single longitude in decimal degrees E, Default: NULL}

\item{latitude}{Single latitude in decimal degrees N, Default: NULL}

\item{radius}{Radius in meters, Default: NULL}

\item{stateDataset}{Name of spatial dataset to use for determining state
codes, Default: 'NaturalEarthAdm1'}

\item{verbose}{Logical controlling the generation of progress messages.}
}
\value{
Updated tibble of known locations.
}
\description{
Incoming \code{longitude} and \code{latitude} values are compared 
against the incoming \code{locationTbl} to see if the are already within
\code{radius} meters of an existing entry.  A new record is created for
if the location is not already found in \code{locationTbl}.
}
\seealso{
\code{\link{addLocations}}
}
