% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaLocationUtils.R
\docType{data}
\name{coreMetadataNames}
\alias{coreMetadataNames}
\title{Names of standard spatial metadata columns}
\format{
A vector with 3 elements
}
\usage{
coreMetadataNames
}
\description{
Character string identifiers of the minimum set of fields
required for a table to be considered a valid "known locations" table.
}
\details{
coreMetadataNames
}
\keyword{datasets}
