% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLocationTable.R
\name{createLocationTable}
\alias{createLocationTable}
\title{Create an empty "known location" table}
\usage{
createLocationTable(collectionName = NULL,
  metadataNames = c("countryCode", "stateCode", "timezone"))
}
\arguments{
\item{collectionName}{Character identifier for this table, Default: NULL}

\item{metadataNames}{Character names of supported spatial metadata, Default: c("countryCode", "stateCode", "timezone")}
}
\value{
Empty "known location" tibble with the specified metadata columns.
}
\description{
Creates an empty "known location" tibble with the following columns:
\itemize{
\item{locationID}
\item{longitude}
\item{latitude}
}
and whatever extra \code{metadataNames} are passed in.
}
\details{
TODO
}
\examples{
\dontrun{
emptyTbl <- createLocationTable(
  "myLocations",
  metadataNames = c("countryCode", "stateCode", "elevation")
)
}
}
