% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaLocationUtils.R
\name{getLocationDataDir}
\alias{getLocationDataDir}
\title{Get location data directory}
\usage{
getLocationDataDir()
}
\value{
Absolute path string.
}
\description{
Returns the directory path where known location data tables are located.
}
\seealso{
\link{LocationDataDir}

\link{setLocationDataDir}
}
\keyword{environment}
