% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLocations.R
\name{getLocations}
\alias{getLocations}
\title{Returns known locations}
\usage{
getLocations(locationTbl = NULL, longitude = NULL, latitude = NULL,
  radius = NULL, nearestIfMultiple = FALSE)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{longitude}{Vector of longitudes in decimal degrees E, Default: NULL}

\item{latitude}{Vector of latitudes in decimal degrees N, Default: NULL}

\item{radius}{Radius in meters, Default: NULL}

\item{nearestIfMultiple}{Logical specifying whether to return the nearest
single location if multiple locations are found within \code{radius}, 
Default: FALSE}
}
\value{
Tibble of known locations
}
\description{
Returns a subset tibble with all locations found in 
\code{locationTbl} that are within \code{radius} meters of the incoming 
locations.
}
