% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNearestLocation.R
\name{getNearestLocation}
\alias{getNearestLocation}
\title{Returns known locations}
\usage{
getNearestLocation(locationTbl = NULL, longitude = NULL,
  latitude = NULL, radius = NULL)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{longitude}{Vector of longitudes in decimal degrees E, Default: NULL}

\item{latitude}{Vector of latitudes in decimal degrees N, Default: NULL}

\item{radius}{Radius in meters, Default: NULL}
}
\value{
Tibble of known locations.
}
\description{
Returns a tibble of known locations from \code{locationTbl}, one 
for each incoming location. If no known location is found for a particular
incoming location, that record in the tibble will contain all \code{NA}.
}
