% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeLocationTable.R
\name{initializeLocationTable}
\alias{initializeLocationTable}
\title{Create an empty known location table}
\usage{
initializeLocationTable(spatialMetadata = NULL)
}
\arguments{
\item{spatialMetadata}{Vector of character names of supported spatial 
metadata, Default: NULL
metadata, Default: NULL}
}
\value{
Empty known location tibble with the specified metadata columns.
}
\description{
Creates an empty known location tibble with the following 
columns of core metadata:
\itemize{
\item{locationID}
\item{locationName}
\item{longitude}
\item{latitude}
\item{elevation}
\item{countryCode}
\item{stateCode}
\item{county}
\item{timezone}
\item{houseNumber}
\item{street}
\item{city}
\item{zip}
}

and whatever extra \code{spatialMetadata} are passed in.
}
\details{
TODO
}
\examples{
\dontrun{
emptyTbl <- initializeLocationTable(
  "myLocations",
  spatialMetadata = c("USCensusCounties")
)
}
}
