% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLocationTable.R
\name{loadLocationTable}
\alias{loadLocationTable}
\title{loadLocationTbl}
\usage{
loadLocationTable(collectionName = NULL)
}
\arguments{
\item{collectionName}{Character identifier for this table, Default: NULL}
}
\value{
Tibble of "known location"s.
}
\description{
Load a "known location" table from the preferred directory.
}
\details{
TODO
}
\seealso{
\code{\link{setLocationDataDir}}
}
