% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_getSingleAddress_TexasAM.R
\name{location_getSingleAddress_TexasAM}
\alias{location_getSingleAddress_TexasAM}
\title{Get an address from a Texas A&M web service}
\usage{
location_getSingleAddress_TexasAM(longitude = NULL, latitude = NULL,
  verbose = TRUE, apiKey = NULL)
}
\arguments{
\item{longitude}{Single longitude in decimal degrees E, Default: NULL}

\item{latitude}{Single latitude in decimal degrees N, Default: NULL}

\item{verbose}{Logical controlling the generation of progress messages.}

\item{apiKey}{Texas A&M Geocoding requires an API key. The first 2500 requests
are free. Default: NULL}
}
\value{
Numeric elevation value.
}
\description{
Texas A&M APIs are used to determine the address associated with
the \code{longitude} and \code{latitude}.
}
\examples{
\donttest{
# Wenatchee
lon <- -120.325278
lat <- 47.423333
apiKey <- apiKey
location_getSingleAddress_TexasAM(lon, lat, apiKey)
}
}
\references{
\url{https://geoservices.tamu.edu/Services/ReverseGeocoding/WebService/v04_01/HTTP.aspx}
}
