% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaLocationUtils.R
\docType{data}
\name{LocationDataDir}
\alias{LocationDataDir}
\title{Directory for location data}
\format{
Absolute path string.
}
\description{
This package maintains an internal directory path which
users can set using \code{setLocationDataDir()}. All package functions use
this directory whenever known location tables are accessed.

The default setting when the package is loaded is \code{getwd()}.
}
\seealso{
\link{getLocationDataDir}

\link{setLocationDataDir}
}
\keyword{environment}
