% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazama_initialize.R
\name{mazama_initialize}
\alias{mazama_initialize}
\title{Initialize with MazamaScience standard directories}
\usage{
mazama_initialize()
}
\value{
No return value.
}
\description{
Convenience function to initialize spatial data. Wraps the 
following setup lines:

\preformatted{
MazamaSpatialUtils::setSpatialDataDir("~/Data/Spatial")

MazamaSpatialUtils::loadSpatialData("EEZCountries")
MazamaSpatialUtils::loadSpatialData("OSMTimezones")
MazamaSpatialUtils::loadSpatialData("NaturalEarthAdm1")
MazamaSpatialUtils::loadSpatialData("USCensusCounties")
}
}
