% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLocation.R
\name{removeLocation}
\alias{removeLocation}
\title{Removes location records from a table}
\usage{
removeLocation(locationTbl = NULL, locationID = NULL, verbose = TRUE)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{locationID}{Vector \code{locationID} strings, Default: NULL}

\item{verbose}{Logical controlling the generation of progress messages.}
}
\value{
Updated tibble of known locations.
}
\description{
Incoming \code{locationID} values are compared 
against the incoming \code{locationTbl} and any matches are removed.
}
\seealso{
\link{addLocation}
}
