% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_export.R
\name{table_export}
\alias{table_export}
\title{Export a known location table}
\usage{
table_export(locationTbl = NULL, outputType = "csv")
}
\arguments{
\item{locationTbl}{Tibble of known locations.}

\item{outputType}{Output format, Default: 'csv'}
}
\value{
Representation of a known location table in the desired format.
}
\description{
Export a known location tibble as CSV format.
}
\examples{
library(MazamaLocationUtils)

locationTbl <- get(data("wa_monitors_500"))
csvString <- table_export(locationTbl)
}
