% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_findOverlappingDistances.R
\name{table_findOverlappingDistances}
\alias{table_findOverlappingDistances}
\title{Find overlapping distances in a known locations table.}
\usage{
table_findOverlappingDistances(
  locationTbl = NULL,
  radius = NULL,
  measure = "geodesic"
)
}
\arguments{
\item{locationTbl}{Tibble of known locations.}

\item{radius}{Radius in meters. Locations are guaranteed to be separatde by 2*R.}

\item{measure}{One of "haversine" "vincenty", "geodesic", or "cheap" 
specifying desired method of geodesic distance calculation. See \code{?geodist::geodist}.}
}
\value{
Tibble of row indices and distances for those locations which overlap.
}
\description{
Calculate distances between all locations within a known
locations table and returns a tibble with the row indices and separation 
distances of those records with overlapping locations.

It is useful when working with new metadata tables to identify overlapping
locations early on so that decisions can be made about the apporpriateness
of the specified \code{radius}.
}
\examples{
library(MazamaLocationUtils)

meta <- wa_airfire_meta

# Anything locations closer than 2 km? (diameter = 2*radius)
table_findOverlappingDistances(meta, radius = 1000)

# How about 4 km?
table_findOverlappingDistances(meta, radius = 2000)


}
