% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_initialize.R
\name{table_initialize}
\alias{table_initialize}
\title{Create an empty known location table}
\usage{
table_initialize()
}
\value{
Empty known location tibble with the specified metadata columns.
}
\description{
Creates an empty known location tibble with the following
columns of core metadata:

\itemize{
\item{locationID}
\item{locationName}
\item{longitude}
\item{latitude}
\item{elevation}
\item{countryCode}
\item{stateCode}
\item{countyName}
\item{timezone}
\item{houseNumber}
\item{street}
\item{city}
\item{zip}
}
}
\examples{
library(MazamaLocationUtils)

# Create an empty Tbl
emptyTbl <- table_initialize()
dplyr::glimpse(emptyTbl)

}
