% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_leafletAdd.R
\name{table_leafletAdd}
\alias{table_leafletAdd}
\title{Add to a leaflet interactive map for known locations}
\usage{
table_leafletAdd(
  map = NULL,
  locationTbl = NULL,
  extraVars = NULL,
  locationOnly = FALSE,
  ...
)
}
\arguments{
\item{map}{Leaflet map.}

\item{locationTbl}{Tibble of known locations.}

\item{extraVars}{Character vector of addition \code{locationTbl} column names
to be shown in leaflet popups.}

\item{locationOnly}{Logical specifying whether to check for all standard
columns.}

\item{...}{Additional arguments passed to \code{leaflet::addCircleMarker()}.}
}
\value{
A leaflet "plot" object which, if not assigned, is rendered in
Rstudio's 'Viewer' tab.
}
\description{
This function adds interactive maps that will be displayed in
  RStudio's 'Viewer' tab.
}
