% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_load.R
\name{table_load}
\alias{table_load}
\title{Locat a known location table}
\usage{
table_load(collectionName = NULL)
}
\arguments{
\item{collectionName}{Character identifier for this table, Default: NULL}
}
\value{
Tibble of known locations.
}
\description{
Load a tibble of known locations from the preferred directory.
}
\details{
TODO
}
\seealso{
\code{\link{setLocationDataDir}}
}
