% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_removeRecord.R
\name{table_removeRecord}
\alias{table_removeRecord}
\title{Remove location records from a table}
\usage{
table_removeRecord(locationTbl = NULL, locationID = NULL,
  verbose = TRUE)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{locationID}{Vector of \code{locationID} strings, Default: NULL}

\item{verbose}{Logical controlling the generation of progress messages.}
}
\value{
Updated tibble of known locations.
}
\description{
Incoming \code{locationID} values are compared 
against the incoming \code{locationTbl} and any matches are removed.
}
\seealso{
\link{table_addLocation}

\link{table_addSingleLocation}

\link{table_updateSingleRecord}
}
