% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_updateColumn.R
\name{table_updateColumn}
\alias{table_updateColumn}
\title{updates a column of metadata in a table}
\usage{
table_updateColumn(locationTbl = NULL, columnName = NULL,
  locationID = NULL, locationData = NULL, verbose = TRUE)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{columnName}{Name to use for the new column, Default: NULL}

\item{locationID}{Vector of \code{locationID} strings, Default: NULL}

\item{locationData}{Vector of data to used at matching records, Default: NULL}

\item{verbose}{Logical controlling the generation of progress messages.}
}
\value{
Updated tibble of known locations.
}
\description{
For matching \code{locationID} records the assaociated 
\code{locatioData} is used to replace any existing value in \code{columnName}.
}
\seealso{
\link{table_addColumn}

\link{table_removeColumn}
}
