% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateLocationTbl}
\alias{validateLocationTbl}
\title{Validate a location table}
\usage{
validateLocationTbl(locationTbl = NULL, locationOnly = TRUE)
}
\arguments{
\item{locationTbl}{Tibble of known locations.}

\item{locationOnly}{Logical specifying whether to check for all standard
columns.}
}
\value{
Invisibly returns \code{TRUE} if no error message has been generated.
}
\description{
Ensures that the incoming table has numeric \code{longitude} and
\code{latitude} columns.
}
