% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateLonsLats}
\alias{validateLonsLats}
\title{Validate longitude and latitude vectors}
\usage{
validateLonsLats(longitude = NULL, latitude = NULL, na.rm = FALSE)
}
\arguments{
\item{longitude}{Vector of longitudes in decimal degrees E.}

\item{latitude}{Vector of latitudes in decimal degrees N.}

\item{na.rm}{Logical specifying whether to remove NA values.}
}
\description{
See \code{\link[MazamaCoreUtils:validateLonsLats]{MazamaCoreUtils::validateLonsLats()}} for details.
}
