% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateMazamaSpatialUtils}
\alias{validateMazamaSpatialUtils}
\title{Validate proper setup of MazamaSpatialUtils}
\usage{
validateMazamaSpatialUtils()
}
\value{
Invisibly returns \code{TRUE} if no error message has been generated.
}
\description{
The \pkg{MazamaSpatialUtils} package mus be properly installed
and initialized before using functions from the \pkg{MazamaLocationUtils} 
package. Functions can test for this
}
