% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaLocationUtils.R
\docType{data}
\name{validMetadataNames}
\alias{validMetadataNames}
\title{Names of standard spatial metadata columns}
\format{A vector with 3 elements}
\usage{
validMetadataNames
}
\description{
Character string identifiers of the different types of spatial 
metadata this package can generate.
}
\details{
validMetadataNames
}
\keyword{datasets}
