% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaLocationUtils.R
\docType{data}
\name{apiKeys}
\alias{apiKeys}
\title{API keys for reverse geocoding services.}
\format{
List of character strings.
}
\description{
This package maintains an internal set of API keys which 
users can set using \code{setAPIKey()}. These keys will be remembered for
the duration of an R session. The following service providers are supported:

\itemize{
\item{\code{"google"}}
\item{\code{"bing"}}
\item{\code{"esri"}}
\item{\code{"texasam"}}
\item{\code{"custom1"}}
\item{\code{"custom2"}}
}
}
\seealso{
\link{getAPIKey}

\link{setAPIKey}
}
\keyword{environment}
