% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createLocationTable.R
\name{createLocationTable}
\alias{createLocationTable}
\title{Create an empty "known location" table}
\usage{
createLocationTable(spatialMetadata = NULL)
}
\arguments{
\item{spatialMetadata}{Vector of character names of supported spatial 
metadata, Default: NULL
metadata, Default: NULL}
}
\value{
Empty "known location" tibble with the specified metadata columns.
}
\description{
Creates an empty "known location" tibble with the following columns:
\itemize{
\item{locationID}
\item{longitude}
\item{latitude}
\item{elevation}
\item{countryCode}
\item{stateCode}
\item{timezone}
}
and whatever extra \code{spatialMetadata} are passed in.
}
\details{
TODO
}
\examples{
\dontrun{
emptyTbl <- createLocationTable(
  "myLocations",
  spatialMetadata = c("USCensusCounties")
)
}
}
