% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_getSingleElevation_USGS.R
\name{location_getSingleElevation_USGS}
\alias{location_getSingleElevation_USGS}
\title{Get elevation data from a USGS web service}
\usage{
location_getSingleElevation_USGS(longitude = NULL, latitude = NULL,
  verbose = TRUE)
}
\arguments{
\item{longitude}{Single longitude in decimal degrees E, Default: NULL}

\item{latitude}{Single latitude in decimal degrees N, Default: NULL}

\item{verbose}{Logical controlling the generation of progress messages.}
}
\value{
Numeric elevation value.
}
\description{
USGS APIs are used to determine the elevation associated with
the \code{longitude} and \code{latitude}.
}
\examples{
\dontrun{
# Set up standard directories and spatial data
mazama_initialize()

# Wenatchee
lon <- -120.325278
lat <- 47.423333
addressList <- location_getSingleElevation_USGS(lon, lat)
}
}
\references{
\url{https://nationalmap.gov/epqs/}
}
