% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveLocationTable.R
\name{saveLocationTable}
\alias{saveLocationTable}
\title{saveLocationTable}
\usage{
saveLocationTable(locationTbl = NULL, collectionName = NULL)
}
\arguments{
\item{locationTbl}{Tibble of "known locations", Default: NULL}

\item{collectionName}{Character identifier for this table, Default: NULL}
}
\value{
File path of saved file.
}
\description{
Save a "known location" tibble to the preferred directory.
}
\details{
TODO
}
