% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MazamaLocationUtils.R
\name{setLocationDataDir}
\alias{setLocationDataDir}
\title{Set location data directory}
\usage{
setLocationDataDir(dataDir)
}
\arguments{
\item{dataDir}{Directory where location tables are stored.}
}
\value{
Silently returns previous value of the data directory.
}
\description{
Sets the data directory where known location data tables are located.
If the directory does not exist, it will be created.
}
\seealso{
\link{LocationDataDir}

\link{getLocationDataDir}
}
\keyword{environment}
