% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_getNearestLocation.R
\name{table_getNearestLocation}
\alias{table_getNearestLocation}
\title{Return known locations}
\usage{
table_getNearestLocation(locationTbl = NULL, longitude = NULL,
  latitude = NULL, radius = NULL)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{longitude}{Vector of longitudes in decimal degrees E, Default: NULL}

\item{latitude}{Vector of latitudes in decimal degrees N, Default: NULL}

\item{radius}{Radius in meters, Default: NULL}
}
\value{
Tibble of known locations.
}
\description{
Returns a tibble of known locations from \code{locationTbl}, one 
for each incoming location. If no known location is found for a particular
incoming location, that record in the tibble will contain all \code{NA}.
}
\examples{
locationTbl <- get(data("wa_monitors_500"))

# Wenatchee
lon <- -120.325278
lat <- 47.423333

# Too small a radius will not find a match
table_getNearestLocation(locationTbl, lon, lat, radius = 50) \%>\% str()

# Expanding the radius will find one
table_getNearestLocation(locationTbl, lon, lat, radius = 5000) \%>\% str()
}
