% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_removeColumn.R
\name{table_removeColumn}
\alias{table_removeColumn}
\title{Remove a column of metadata in a table}
\usage{
table_removeColumn(locationTbl = NULL, columnName = NULL,
  verbose = TRUE)
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{columnName}{Name of the colun to be removed, Default: NULL}

\item{verbose}{Logical controlling the generation of progress messages.}
}
\value{
Updated tibble of known locations.
}
\description{
Remove the column matching \code{columnName}. This function 
can be used in pipelines.
}
\examples{
\donttest{
# Set up standard directories and spatial data
mazama_initialize()

# Set the directory for saving location tables
setLocationDataDir(tempdir())

locationTbl <- get(data("wa_monitors_500"))

# Save it
table_save(locationTbl, "my_table")

# Load it
my_table <- table_load("my_table")

# Modify and save again (backup of previous will be generated)
my_table \%>\%
  table_addColumn("my_column") \%>\%
  table_save("my_table")
  
# Change your mind
table_load("my_table") \%>\%
  table_removeColumn("my_column") \%>\%
  table_save("my_table")
  
list.files(getLocationDataDir(), pattern = "my_table")
}
}
\seealso{
\link{table_addColumn}

\link{table_removeColumn}
}
