% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_save.R
\name{table_save}
\alias{table_save}
\title{Save a known location table}
\usage{
table_save(locationTbl = NULL, collectionName = NULL, backup = TRUE,
  outputType = "rda")
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{collectionName}{Character identifier for this table, Default: NULL}

\item{backup}{Logical specifying whether to save a backup version of any
existing tables sharing \code{collectionName}.}

\item{outputType}{Output format, Default: 'rda'}
}
\value{
File path of saved file.
}
\description{
Save a tibble of known locations to the preferred directory.
}
\details{
Backup files are saved with "YYYY-mm-ddTHH:MM:SS"
}
\examples{
\donttest{
# Set up standard directories and spatial data
mazama_initialize()

# Set the directory for saving location tables
setLocationDataDir(tempdir())

locationTbl <- get(data("wa_monitors_500"))

# Save it
table_save(locationTbl, "my_table")

# Load it
my_table <- table_load("my_table")

# Modify and save again (backup of previous will be generated)
my_table \%>\%
  table_addColumn("my_column") \%>\%
  table_save("my_table")
  
list.files(getLocationDataDir(), pattern = "my_table")
}
}
