% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_save.R
\name{table_save}
\alias{table_save}
\title{Save a known location table}
\usage{
table_save(locationTbl = NULL, collectionName = NULL, backup = TRUE,
  outputType = "rda")
}
\arguments{
\item{locationTbl}{Tibble of known locations, Default: NULL}

\item{collectionName}{Character identifier for this table, Default: NULL}

\item{backup}{Logical specifying whether to save a backup version of any
existing tables sharing \code{collectionName}.}

\item{outputType}{Output format, Default: 'rda'}
}
\value{
File path of saved file.
}
\description{
Save a tibble of known locations to the preferred directory.
}
\details{
Backup files are saved with "YYYY-mm-ddTHH:MM:SS"
}
