% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateLonLat}
\alias{validateLonLat}
\title{Validate longitude and latitude values}
\usage{
validateLonLat(longitude = NULL, latitude = NULL)
}
\description{
See \code{\link[MazamaCoreUtils:validateLonLat]{MazamaCoreUtils::validateLonLat()}} for details.
}
\keyword{internal}
