% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateLonsLats}
\alias{validateLonsLats}
\title{Validate longitude and latitude vectors}
\usage{
validateLonsLats(longitude = NULL, latitude = NULL, na.rm = FALSE)
}
\description{
See \code{MazamaCoreUtils::\link[MazamaCoreUtils:validateLonsLats]{validateLonsLats}} for details.
}
\keyword{internal}
