% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wa_airfire_meta}
\alias{wa_airfire_meta}
\title{Washington monitor metadata dataset}
\format{
A tibble with 73 rows and 19 columns of data.
}
\usage{
wa_airfire_meta
}
\description{
The \code{wa_pwfsl_meta} dataset provides a set of Washington
state air quality monitor metadata used by the USFS AirFire group.
This dataset was generated on 2021-10-19 by running:

\preformatted{
library(PWFSLSmoke)

wa_airfire_meta <- 
  monitor_loadLatest() %>% 
  monitor_subset(stateCodes = "WA") \%>\%
  monitor_extractMeta()
 
save(wa_airfire_meta, file = "data/wa_airfire_meta.rda")
}
}
\keyword{datasets}
